backend-python/requirements.txt
Flask
flask-cors
google-api-python-client
google-auth
python-dotenv
scikit-learn
pandas
joblib
mysql-connector-python
